requirements.txt
# SV Pipeline Python Dependencies

# ===== Core Dependencies =====
python-dotenv==1.0.0          # Environment variable management

# ===== Web Scraping =====
requests==2.31.0              # HTTP requests
beautifulsoup4==4.12.3        # HTML parsing
lxml==5.1.0                   # XML/HTML parser (faster than html.parser)

# PDF extraction
pypdf==4.0.1                  # PDF reading

# Optional: Advanced scraping (uncomment if needed)
# playwright==1.41.0          # JavaScript rendering
# selenium==4.17.2            # Alternative browser automation

# ===== LLM APIs =====
anthropic==0.18.1             # Claude API client
# openai==1.12.0              # OpenAI API client (alternative)

# ===== Google Sheets Integration =====
gspread==6.0.2                # Google Sheets API wrapper
google-auth==2.27.0           # Google authentication
google-auth-oauthlib==1.2.0   # OAuth support
google-auth-httplib2==0.2.0   # HTTP client for Google Auth

# ===== Data Processing =====
pydantic==2.6.1               # Data validation (for JSON schemas)

# ===== Document Generation =====
jinja2==3.1.3                 # Template rendering
markdown2==2.4.12             # Markdown processing

# Optional: PDF generation (uncomment if needed)
# weasyprint==60.2            # HTML to PDF conversion

# ===== Utilities =====
requests-cache==1.1.1         # HTTP caching (reduce redundant scrapes)
urllib3==2.2.0                # URL utilities
python-dateutil==2.8.2        # Date parsing

# ===== Logging & Monitoring =====
colorlog==6.8.2               # Colored console logs
